+++ packages/template-minima-title/src/Root.tsx
@@ -203,7 +203,7 @@ export const RemotionRoot: React.FC = () => {
         <Composition
           id="StonecrestReveal"
           component={StonecrestReveal}
-          durationInFrames={600}
+          durationInFrames={420}
           fps={30}
           width={1920}
           height={1080}